+++ src/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: jkoers <jkoers@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/27 13:09:27 by jkoers        #+#    #+#                 */
-/*   Updated: 2020/10/28 23:09:40 by jkoers        ########   odam.nl         */
+/*   Updated: 2020/11/02 12:41:29 by jkoers        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char	*new_s;
-	size_t	s_len;
-	size_t	i;
+	char			*new_s;
+	size_t			s_len;
+	unsigned int	i;
 
 	s_len = ft_strlen((char *)s);
 	new_s = malloc((s_len + 1) * sizeof(char));
